rename Capabilities and add parsed capabilities
